backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
google-generativeai
PyPDF2
pandas
python-dotenv
langchain
langchain-text-splitters
sentence-transformers
aiofiles
